% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anno_signature_genes_mouse}
\alias{anno_signature_genes_mouse}
\title{Annotation of signature genes of different cell types of mouse mammary gland.}
\format{
A data frame with 2427 rows and 4 columns:
\describe{
\item{GeneID}{Gene ID from NCBI}
\item{Symbol}{Gene symbol}
\item{Chr}{Chromosome location}
\item{gene_type}{Signature gene type}
}
}
\source{
\url{https://breast-cancer-research.biomedcentral.com/articles/10.1186/s13058-021-01445-4}
}
\usage{
anno_signature_genes_mouse
}
\description{
A dataset containing 1470 Basal genes, 428 LP genes and 528 ML genes
}
\keyword{datasets}
