% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_cutoffs.R
\name{calculate_cv_by_row}
\alias{calculate_cv_by_row}
\title{Calculate the coefficient of variation (CV)}
\usage{
calculate_cv_by_row(cutoff_exp, ...)
}
\arguments{
\item{cutoff_exp}{Expression cut-offs used to count the signature gene number}

\item{...}{Other parameters in generate_data_for_ternary()}
}
\description{
Calculate the coefficient of variation (CV) of each row of
the result matrix from generate_data_for_ternary().
calculate_cv_by_row() is not used separately but used in estimate_optimized_cutoffs()
to obtain the optimized cut-off.
}
\examples{

calculate_cv_by_row(
  data_exp_mat = edgeR::cpm(example_dge_data$counts,
                            log = FALSE),
  anno_signature_genes = anno_signature_genes_mouse,
  gene_name_col = "GeneID",
  gene_type_col = "gene_type",
  weight_by_gene_count = FALSE,
  cutoff_exp = 100,
  prior_count = 0
)


}
