% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_dge_data}
\alias{example_dge_data}
\title{An example dataset used in the example codes.}
\format{
A DGElist object with raw count matrix, sample annotation data.frame and gene annotation data.frame:
\describe{
\item{counts}{raw count matrix}
\item{samples}{Sample annotation}
\item{genes}{Gene annotation}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE63310}
}
\usage{
example_dge_data
}
\description{
A DGElist object containing 3 Basal, 3 LP and 3 ML bulk RNA-seq samples
}
\keyword{datasets}
