% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_seu}
\alias{example_seu}
\title{An example single cell RNA-Seq dataset}
\format{
An object of class \code{Seurat} with 13750 rows and 80 columns.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSM2510617}
}
\usage{
example_seu
}
\description{
A Seurat object containing 13750 genes and 80 cells (20 Basal, 20 LP, 20 ML and 20 Luminal-int cells). This example dataset is a subset of an adult mammary gland single cell RNA-Seq sample from Pal et al. 2017. Nature communications.
}
\keyword{datasets}
