% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_layout_matrix.R
\name{prepare_layout_matrix}
\alias{prepare_layout_matrix}
\title{Generate layout matrix for gridExtra::grid.arrange}
\usage{
prepare_layout_matrix(
  grobs_list = NULL,
  last_is_legend = TRUE,
  n_col = NULL,
  scale_legend = 1
)
}
\arguments{
\item{grobs_list}{A list of graphical objects (grobs)}

\item{last_is_legend}{Whether last graphical object is legend}

\item{n_col}{Number of columns for plotting}

\item{scale_legend}{Scaled width of legend}
}
\description{
Generate layout matrix for gridExtra::grid.arrange
}
\examples{
prepare_layout_matrix( grobs_list = list(A=1,B=2,C=3) )

prepare_layout_matrix( grobs_list = 1:3,
                       last_is_legend = FALSE )

}
